+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix timed out batch and transfer logging to ensure that the log file is closed correctly
 - Reorder log format to bring log level to the front
 - Add retry logic to SSH based connections
+- Prevent local based transfers from using the staging directory
 
 # v24.8.0